docs($Exercise): update JavaDoc

+++ src/main/java/com/jmsoftware/exrxnetcrawlerserver/exercise/domain/ExerciseRelatedMusclePo.java
@@ -7,8 +7,8 @@ import lombok.Data;
  * <p>
  * Exercise Persistence Object.
  *
- * @author Johnny Miller (锺俊), email: johnnysviva@outlook.com
- * @date 1/20/20 4:18 PM
+ * @author Johnny Miller (鍾俊), e-mail: johnnysviva@outlook.com
+ * @date 1/21/20 10:53 AM
  **/
 @Data
 public class ExerciseRelatedMusclePo {
@@ -21,8 +21,20 @@ public class ExerciseRelatedMusclePo {
      */
     private Long muscleId;
     /**
-     * Related muscle type.
-     * 1 - target, 2 - synergists, 3 - stabilizers
+     * <p>Related muscle type. Muscle movement classification.</p>
+     * <ul>
+     * <li>1 - Agonist</li>
+     * <li>2 - Antagonist</li>
+     * <li>3 - Target</li>
+     * <li>4 - Synergist</li>
+     * <li>5 - Stabilizer</li>
+     * <li>6 - Dynamic Stabilizer</li>
+     * <li>7 - Antagonist Stabilizer</li>
+     *
+     * </ul>
+     * <p>'https://exrx.net/Kinesiology/Glossary#MuscleMovClass' target='_blank' class='url'>https://exrx
+     * .net/Kinesiology/Glossary#MuscleMovClass</a>
+     * </p>
      */
-    private Byte relatedMuscleType;
+    private Integer relatedMuscleType;
 }
